+++ apple2js.html
@@ -309,6 +309,6 @@
       </filter>
   </svg>
 
-  <script src="dist/main2.js"></script>
+  <script src="dist/main2.bundle.js"></script>
 </body>
 </html>
