+++ apple2jse.html
@@ -29,10 +29,6 @@
 <link rel="stylesheet" type="text/css" href="css/apple2.css" />
 <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.2/css/all.css" />
 
-<script>
-  window.e = true
-</script>
-
 <!-- Disk Index -->
 <script type="text/javascript" src="json/disks/index.js"></script>
 
@@ -318,6 +314,6 @@
       </filter>
   </svg>
 
-  <script src="dist/main2e.js"></script>
+  <script src="dist/main2e.bundle.js"></script>
 </body>
 </html>
