+++ js/entry2e.js
@@ -1,4 +1,6 @@
 import { apple2 } from './main2e';
 import * as UI from './ui/apple2';
 
+window.e = true;
+
 export const Apple2 = { ...UI, apple2 };
