+++ js/roms/character/apple2enh_char.ts
@@ -1,4 +1,4 @@
-import { ReadonlyUint8Array } from '../types';
+import { ReadonlyUint8Array } from '../../types';
 
 const apple2enh_charset = new Uint8Array([
     0x1c,0x22,0x2a,0x3a,0x1a,0x02,0x3c,0x00,
