+++ js/roms/character/rmfont_char.ts
@@ -1,20 +1,18 @@
-import { ReadonlyUint8Array } from '../types';
+import { ReadonlyUint8Array } from '../../types';
 
-/* exported apple2e_charset */
-
-const apple2e_charset = new Uint8Array([
-    0x1c,0x22,0x2a,0x3a,0x1a,0x02,0x3c,0x00,
+const rmfont_charset = new Uint8Array([
+    0x3c,0x42,0x59,0x55,0x55,0x39,0x02,0x3c,
     0x08,0x14,0x22,0x22,0x3e,0x22,0x22,0x00,
-    0x1e,0x22,0x22,0x1e,0x22,0x22,0x1e,0x00,
+    0x1e,0x24,0x24,0x1c,0x24,0x24,0x1e,0x00,
     0x1c,0x22,0x02,0x02,0x02,0x22,0x1c,0x00,
-    0x1e,0x22,0x22,0x22,0x22,0x22,0x1e,0x00,
+    0x1e,0x24,0x24,0x24,0x24,0x24,0x1e,0x00,
     0x3e,0x02,0x02,0x1e,0x02,0x02,0x3e,0x00,
     0x3e,0x02,0x02,0x1e,0x02,0x02,0x02,0x00,
-    0x3c,0x02,0x02,0x02,0x32,0x22,0x3c,0x00,
+    0x1c,0x22,0x02,0x02,0x32,0x22,0x3c,0x00,
     0x22,0x22,0x22,0x3e,0x22,0x22,0x22,0x00,
     0x1c,0x08,0x08,0x08,0x08,0x08,0x1c,0x00,
-    0x20,0x20,0x20,0x20,0x20,0x22,0x1c,0x00,
-    0x22,0x12,0x0a,0x06,0x0a,0x12,0x22,0x00,
+    0x38,0x10,0x10,0x10,0x10,0x12,0x0c,0x00,
+    0x22,0x12,0x0a,0x16,0x12,0x22,0x22,0x00,
     0x02,0x02,0x02,0x02,0x02,0x02,0x3e,0x00,
     0x22,0x36,0x2a,0x2a,0x22,0x22,0x22,0x00,
     0x22,0x22,0x26,0x2a,0x32,0x22,0x22,0x00,
@@ -25,7 +23,7 @@ const apple2e_charset = new Uint8Array([
     0x1c,0x22,0x02,0x1c,0x20,0x22,0x1c,0x00,
     0x3e,0x08,0x08,0x08,0x08,0x08,0x08,0x00,
     0x22,0x22,0x22,0x22,0x22,0x22,0x1c,0x00,
-    0x22,0x22,0x22,0x22,0x22,0x14,0x08,0x00,
+    0x22,0x22,0x22,0x22,0x14,0x14,0x08,0x00,
     0x22,0x22,0x22,0x2a,0x2a,0x36,0x22,0x00,
     0x22,0x22,0x14,0x08,0x14,0x22,0x22,0x00,
     0x22,0x22,0x14,0x08,0x08,0x08,0x08,0x00,
@@ -36,69 +34,69 @@ const apple2e_charset = new Uint8Array([
     0x00,0x00,0x08,0x14,0x22,0x00,0x00,0x00,
     0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7f,
     0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x08,0x08,0x08,0x08,0x08,0x00,0x08,0x00,
-    0x14,0x14,0x14,0x00,0x00,0x00,0x00,0x00,
+    0x08,0x1c,0x1c,0x08,0x08,0x00,0x08,0x00,
+    0x36,0x36,0x14,0x00,0x00,0x00,0x00,0x00,
     0x14,0x14,0x3e,0x14,0x3e,0x14,0x14,0x00,
     0x08,0x3c,0x0a,0x1c,0x28,0x1e,0x08,0x00,
     0x06,0x26,0x10,0x08,0x04,0x32,0x30,0x00,
     0x04,0x0a,0x0a,0x04,0x2a,0x12,0x2c,0x00,
-    0x08,0x08,0x08,0x00,0x00,0x00,0x00,0x00,
+    0x18,0x10,0x08,0x00,0x00,0x00,0x00,0x00,
     0x08,0x04,0x02,0x02,0x02,0x04,0x08,0x00,
     0x08,0x10,0x20,0x20,0x20,0x10,0x08,0x00,
-    0x08,0x2a,0x1c,0x08,0x1c,0x2a,0x08,0x00,
+    0x08,0x2a,0x1c,0x3e,0x1c,0x2a,0x08,0x00,
     0x00,0x08,0x08,0x3e,0x08,0x08,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x08,0x08,0x04,0x00,
+    0x00,0x00,0x00,0x00,0x00,0x08,0x08,0x04,
     0x00,0x00,0x00,0x3e,0x00,0x00,0x00,0x00,
     0x00,0x00,0x00,0x00,0x00,0x00,0x08,0x00,
     0x00,0x20,0x10,0x08,0x04,0x02,0x00,0x00,
-    0x1c,0x22,0x32,0x2a,0x26,0x22,0x1c,0x00,
+    0x1c,0x22,0x22,0x2a,0x22,0x22,0x1c,0x00,
     0x08,0x0c,0x08,0x08,0x08,0x08,0x1c,0x00,
     0x1c,0x22,0x20,0x18,0x04,0x02,0x3e,0x00,
     0x3e,0x20,0x10,0x18,0x20,0x22,0x1c,0x00,
     0x10,0x18,0x14,0x12,0x3e,0x10,0x10,0x00,
-    0x3e,0x02,0x1e,0x20,0x20,0x22,0x1c,0x00,
+    0x3e,0x02,0x1e,0x22,0x20,0x22,0x1c,0x00,
     0x38,0x04,0x02,0x1e,0x22,0x22,0x1c,0x00,
     0x3e,0x20,0x10,0x08,0x04,0x04,0x04,0x00,
     0x1c,0x22,0x22,0x1c,0x22,0x22,0x1c,0x00,
-    0x1c,0x22,0x22,0x3c,0x20,0x10,0x0e,0x00,
+    0x1c,0x22,0x22,0x3c,0x20,0x10,0x08,0x00,
     0x00,0x00,0x08,0x00,0x08,0x00,0x00,0x00,
-    0x00,0x00,0x08,0x00,0x08,0x08,0x04,0x00,
+    0x00,0x00,0x08,0x00,0x00,0x08,0x08,0x04,
     0x10,0x08,0x04,0x02,0x04,0x08,0x10,0x00,
     0x00,0x00,0x3e,0x00,0x3e,0x00,0x00,0x00,
     0x04,0x08,0x10,0x20,0x10,0x08,0x04,0x00,
-    0x1c,0x22,0x10,0x08,0x08,0x00,0x08,0x00,
-    0x1c,0x22,0x2a,0x3a,0x1a,0x02,0x3c,0x00,
-    0x08,0x14,0x22,0x22,0x3e,0x22,0x22,0x00,
-    0x1e,0x22,0x22,0x1e,0x22,0x22,0x1e,0x00,
-    0x1c,0x22,0x02,0x02,0x02,0x22,0x1c,0x00,
-    0x1e,0x22,0x22,0x22,0x22,0x22,0x1e,0x00,
-    0x3e,0x02,0x02,0x1e,0x02,0x02,0x3e,0x00,
-    0x3e,0x02,0x02,0x1e,0x02,0x02,0x02,0x00,
-    0x3c,0x02,0x02,0x02,0x32,0x22,0x3c,0x00,
-    0x22,0x22,0x22,0x3e,0x22,0x22,0x22,0x00,
-    0x1c,0x08,0x08,0x08,0x08,0x08,0x1c,0x00,
-    0x20,0x20,0x20,0x20,0x20,0x22,0x1c,0x00,
-    0x22,0x12,0x0a,0x06,0x0a,0x12,0x22,0x00,
-    0x02,0x02,0x02,0x02,0x02,0x02,0x3e,0x00,
-    0x22,0x36,0x2a,0x2a,0x22,0x22,0x22,0x00,
-    0x22,0x22,0x26,0x2a,0x32,0x22,0x22,0x00,
-    0x1c,0x22,0x22,0x22,0x22,0x22,0x1c,0x00,
-    0x1e,0x22,0x22,0x1e,0x02,0x02,0x02,0x00,
-    0x1c,0x22,0x22,0x22,0x2a,0x12,0x2c,0x00,
-    0x1e,0x22,0x22,0x1e,0x0a,0x12,0x22,0x00,
-    0x1c,0x22,0x02,0x1c,0x20,0x22,0x1c,0x00,
-    0x3e,0x08,0x08,0x08,0x08,0x08,0x08,0x00,
-    0x22,0x22,0x22,0x22,0x22,0x22,0x1c,0x00,
-    0x22,0x22,0x22,0x22,0x22,0x14,0x08,0x00,
-    0x22,0x22,0x22,0x2a,0x2a,0x36,0x22,0x00,
-    0x22,0x22,0x14,0x08,0x14,0x22,0x22,0x00,
-    0x22,0x22,0x14,0x08,0x08,0x08,0x08,0x00,
-    0x3e,0x20,0x10,0x08,0x04,0x02,0x3e,0x00,
-    0x3e,0x06,0x06,0x06,0x06,0x06,0x3e,0x00,
-    0x00,0x02,0x04,0x08,0x10,0x20,0x00,0x00,
-    0x3e,0x30,0x30,0x30,0x30,0x30,0x3e,0x00,
-    0x00,0x00,0x08,0x14,0x22,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x7f,
+    0x1c,0x22,0x20,0x10,0x08,0x00,0x08,0x00,
+    0xef,0xf7,0xc9,0x80,0xc0,0xc0,0x81,0xc9,
+    0xef,0xf7,0xc9,0xbe,0xde,0xde,0xb5,0xc9,
+    0xff,0xff,0xfd,0xf9,0xf1,0xe1,0xc9,0xbd,
+    0x80,0xdd,0xeb,0xf7,0xf7,0xeb,0xd5,0x80,
+    0xff,0xbf,0xdf,0xee,0xf5,0xfb,0xfb,0xff,
+    0x80,0xc0,0xa0,0x93,0x8a,0x84,0x84,0x80,
+    0x40,0x40,0x40,0x44,0x46,0x7f,0x06,0x04,
+    0x00,0x7f,0x00,0x7f,0x00,0x7f,0x7f,0x00,
+    0xf7,0xfb,0xfd,0x80,0xfd,0xfb,0xf7,0xff,
+    0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xd5,
+    0xf7,0xf7,0xf7,0xf7,0xb6,0xd5,0xe3,0xf7,
+    0xf7,0xe3,0xd5,0xb6,0xf7,0xf7,0xf7,0xf7,
+    0x80,0xff,0xff,0xff,0xff,0xff,0xff,0xff,
+    0xbf,0xbf,0xbf,0xbb,0xb9,0x80,0xf9,0xfb,
+    0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,0xc0,
+    0xec,0xe7,0xe3,0x81,0xe3,0xe7,0xef,0x90,
+    0x9b,0xf3,0xe3,0xc0,0xe3,0xf3,0xfb,0x84,
+    0xbf,0xb7,0xf7,0x80,0xc1,0xe3,0xb7,0xbf,
+    0xbf,0xb7,0xe3,0xc1,0x80,0xf7,0xb7,0xbf,
+    0xff,0xff,0xff,0x80,0xff,0xff,0xff,0xff,
+    0xfe,0xfe,0xfe,0xfe,0xfe,0xfe,0xfe,0x80,
+    0xf7,0xef,0xdf,0x80,0xdf,0xef,0xf7,0xff,
+    0xd5,0xaa,0xd5,0xaa,0xd5,0xaa,0xd5,0xaa,
+    0xaa,0xd5,0xaa,0xd5,0xaa,0xd5,0xaa,0xd5,
+    0xff,0xc1,0xbe,0xfe,0xfe,0xfe,0x80,0xff,
+    0xff,0xff,0xc0,0xbf,0xbf,0xbf,0x80,0xff,
+    0xbf,0xbf,0xbf,0xbf,0xbf,0xbf,0xbf,0xbf,
+    0xf7,0xe3,0xc1,0x80,0xc1,0xe3,0xf7,0xff,
+    0x80,0xff,0xff,0xff,0xff,0xff,0xff,0x80,
+    0xeb,0xeb,0x88,0xff,0x88,0xeb,0xeb,0xff,
+    0x80,0xbf,0xbf,0xb3,0xb3,0xbf,0xbf,0x80,
+    0xfe,0xfe,0xfe,0xfe,0xfe,0xfe,0xfe,0xfe,
     0x04,0x08,0x10,0x00,0x00,0x00,0x00,0x00,
     0x00,0x00,0x1c,0x20,0x3c,0x22,0x3c,0x00,
     0x02,0x02,0x1e,0x22,0x22,0x22,0x1e,0x00,
@@ -116,13 +114,13 @@ const apple2e_charset = new Uint8Array([
     0x00,0x00,0x1e,0x22,0x22,0x22,0x22,0x00,
     0x00,0x00,0x1c,0x22,0x22,0x22,0x1c,0x00,
     0x00,0x00,0x1e,0x22,0x22,0x1e,0x02,0x02,
-    0x00,0x00,0x3c,0x22,0x22,0x3c,0x20,0x20,
-    0x00,0x00,0x3a,0x06,0x02,0x02,0x02,0x00,
+    0x00,0x00,0x3c,0x22,0x22,0x3c,0x20,0x60,
+    0x00,0x00,0x1a,0x26,0x02,0x02,0x02,0x00,
     0x00,0x00,0x3c,0x02,0x1c,0x20,0x1e,0x00,
     0x04,0x04,0x1e,0x04,0x04,0x24,0x18,0x00,
     0x00,0x00,0x22,0x22,0x22,0x32,0x2c,0x00,
     0x00,0x00,0x22,0x22,0x22,0x14,0x08,0x00,
-    0x00,0x00,0x22,0x22,0x2a,0x2a,0x36,0x00,
+    0x00,0x00,0x22,0x22,0x2a,0x2a,0x14,0x00,
     0x00,0x00,0x22,0x14,0x08,0x14,0x22,0x00,
     0x00,0x00,0x22,0x22,0x22,0x3c,0x20,0x1c,
     0x00,0x00,0x3e,0x10,0x08,0x04,0x3e,0x00,
@@ -130,19 +128,19 @@ const apple2e_charset = new Uint8Array([
     0x08,0x08,0x08,0x08,0x08,0x08,0x08,0x08,
     0x0e,0x18,0x18,0x30,0x18,0x18,0x0e,0x00,
     0x2c,0x1a,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x2a,0x14,0x2a,0x14,0x2a,0x00,0x00,
-    0xe3,0xdd,0xd5,0xc5,0xe5,0xfd,0xc3,0xff,
+    0x6f,0x77,0x41,0x00,0x40,0x40,0x00,0x49,
+    0x43,0x3c,0x26,0x2a,0x2a,0x46,0x7d,0x43,
     0xf7,0xeb,0xdd,0xdd,0xc1,0xdd,0xdd,0xff,
-    0xe1,0xdd,0xdd,0xe1,0xdd,0xdd,0xe1,0xff,
+    0xe1,0xdb,0xdb,0xe3,0xdb,0xdb,0xe1,0x7f,
     0xe3,0xdd,0xfd,0xfd,0xfd,0xdd,0xe3,0xff,
-    0xe1,0xdd,0xdd,0xdd,0xdd,0xdd,0xe1,0xff,
+    0xe1,0xdb,0xdb,0xdb,0xdb,0xdb,0xe1,0x7f,
     0xc1,0xfd,0xfd,0xe1,0xfd,0xfd,0xc1,0xff,
     0xc1,0xfd,0xfd,0xe1,0xfd,0xfd,0xfd,0xff,
-    0xc3,0xfd,0xfd,0xfd,0xcd,0xdd,0xc3,0xff,
+    0xe3,0xdd,0xfd,0xfd,0xcd,0xdd,0xc3,0x7f,
     0xdd,0xdd,0xdd,0xc1,0xdd,0xdd,0xdd,0xff,
     0xe3,0xf7,0xf7,0xf7,0xf7,0xf7,0xe3,0xff,
-    0xdf,0xdf,0xdf,0xdf,0xdf,0xdd,0xe3,0xff,
-    0xdd,0xed,0xf5,0xf9,0xf5,0xed,0xdd,0xff,
+    0xc7,0xef,0xef,0xef,0xef,0xed,0xf3,0x7f,
+    0xdd,0xed,0xf5,0xe9,0xed,0xdd,0xdd,0x7f,
     0xfd,0xfd,0xfd,0xfd,0xfd,0xfd,0xc1,0xff,
     0xdd,0xc9,0xd5,0xd5,0xdd,0xdd,0xdd,0xff,
     0xdd,0xdd,0xd9,0xd5,0xcd,0xdd,0xdd,0xff,
@@ -153,7 +151,7 @@ const apple2e_charset = new Uint8Array([
     0xe3,0xdd,0xfd,0xe3,0xdf,0xdd,0xe3,0xff,
     0xc1,0xf7,0xf7,0xf7,0xf7,0xf7,0xf7,0xff,
     0xdd,0xdd,0xdd,0xdd,0xdd,0xdd,0xe3,0xff,
-    0xdd,0xdd,0xdd,0xdd,0xdd,0xeb,0xf7,0xff,
+    0xdd,0xdd,0xdd,0xdd,0xeb,0xeb,0xf7,0x7f,
     0xdd,0xdd,0xdd,0xd5,0xd5,0xc9,0xdd,0xff,
     0xdd,0xdd,0xeb,0xf7,0xeb,0xdd,0xdd,0xff,
     0xdd,0xdd,0xeb,0xf7,0xf7,0xf7,0xf7,0xff,
@@ -164,49 +162,49 @@ const apple2e_charset = new Uint8Array([
     0xff,0xff,0xf7,0xeb,0xdd,0xff,0xff,0xff,
     0xff,0xff,0xff,0xff,0xff,0xff,0xff,0x80,
     0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,
-    0xf7,0xf7,0xf7,0xf7,0xf7,0xff,0xf7,0xff,
-    0xeb,0xeb,0xeb,0xff,0xff,0xff,0xff,0xff,
+    0x77,0x63,0x63,0x77,0x77,0x7f,0x77,0x7f,
+    0x49,0x49,0x6b,0x7f,0x7f,0x7f,0x7f,0x7f,
     0xeb,0xeb,0xc1,0xeb,0xc1,0xeb,0xeb,0xff,
     0xf7,0xc3,0xf5,0xe3,0xd7,0xe1,0xf7,0xff,
     0xf9,0xd9,0xef,0xf7,0xfb,0xcd,0xcf,0xff,
     0xfb,0xf5,0xf5,0xfb,0xd5,0xed,0xd3,0xff,
-    0xf7,0xf7,0xf7,0xff,0xff,0xff,0xff,0xff,
+    0x67,0x6f,0x77,0x7f,0x7f,0x7f,0x7f,0x7f,
     0xf7,0xfb,0xfd,0xfd,0xfd,0xfb,0xf7,0xff,
     0xf7,0xef,0xdf,0xdf,0xdf,0xef,0xf7,0xff,
-    0xf7,0xd5,0xe3,0xf7,0xe3,0xd5,0xf7,0xff,
+    0xf7,0xd5,0xe3,0xc1,0xe3,0xd5,0xf7,0xff,
     0xff,0xf7,0xf7,0xc1,0xf7,0xf7,0xff,0xff,
-    0xff,0xff,0xff,0xff,0xf7,0xf7,0xfb,0xff,
+    0xff,0xff,0xff,0xff,0xff,0xf7,0xf7,0xfb,
     0xff,0xff,0xff,0xc1,0xff,0xff,0xff,0xff,
     0xff,0xff,0xff,0xff,0xff,0xff,0xf7,0xff,
     0xff,0xdf,0xef,0xf7,0xfb,0xfd,0xff,0xff,
-    0xe3,0xdd,0xcd,0xd5,0xd9,0xdd,0xe3,0xff,
+    0x63,0x5d,0x5d,0x55,0x5d,0x5d,0x63,0x7f,
     0xf7,0xf3,0xf7,0xf7,0xf7,0xf7,0xe3,0xff,
     0xe3,0xdd,0xdf,0xe7,0xfb,0xfd,0xc1,0xff,
     0xc1,0xdf,0xef,0xe7,0xdf,0xdd,0xe3,0xff,
     0xef,0xe7,0xeb,0xed,0xc1,0xef,0xef,0xff,
-    0xc1,0xfd,0xe1,0xdf,0xdf,0xdd,0xe3,0xff,
+    0x41,0x7d,0x61,0x5d,0x5f,0x5d,0x63,0x7f,
     0xc7,0xfb,0xfd,0xe1,0xdd,0xdd,0xe3,0xff,
     0xc1,0xdf,0xef,0xf7,0xfb,0xfb,0xfb,0xff,
     0xe3,0xdd,0xdd,0xe3,0xdd,0xdd,0xe3,0xff,
-    0xe3,0xdd,0xdd,0xc3,0xdf,0xef,0xf1,0xff,
+    0xe3,0xdd,0xdd,0xc3,0xdf,0xef,0xf7,0xff,
     0xff,0xff,0xf7,0xff,0xf7,0xff,0xff,0xff,
-    0xff,0xff,0xf7,0xff,0xf7,0xf7,0xfb,0xff,
+    0xff,0xff,0xf7,0xff,0xff,0xf7,0xf7,0xfb,
     0xef,0xf7,0xfb,0xfd,0xfb,0xf7,0xef,0xff,
     0xff,0xff,0xc1,0xff,0xc1,0xff,0xff,0xff,
     0xfb,0xf7,0xef,0xdf,0xef,0xf7,0xfb,0xff,
-    0xe3,0xdd,0xef,0xf7,0xf7,0xff,0xf7,0xff,
-    0xe3,0xdd,0xd5,0xc5,0xe5,0xfd,0xc3,0xff,
+    0x63,0x5d,0x5f,0x6f,0x77,0x7f,0x77,0x7f,
+    0x43,0x3c,0x26,0x2a,0x2a,0x46,0x7d,0x43,
     0xf7,0xeb,0xdd,0xdd,0xc1,0xdd,0xdd,0xff,
-    0xe1,0xdd,0xdd,0xe1,0xdd,0xdd,0xe1,0xff,
+    0xe1,0xdb,0xdb,0xe3,0xdb,0xdb,0xe1,0xff,
     0xe3,0xdd,0xfd,0xfd,0xfd,0xdd,0xe3,0xff,
-    0xe1,0xdd,0xdd,0xdd,0xdd,0xdd,0xe1,0xff,
+    0xe1,0xdb,0xdb,0xdb,0xdb,0xdb,0xe1,0xff,
     0xc1,0xfd,0xfd,0xe1,0xfd,0xfd,0xc1,0xff,
     0xc1,0xfd,0xfd,0xe1,0xfd,0xfd,0xfd,0xff,
-    0xc3,0xfd,0xfd,0xfd,0xcd,0xdd,0xc3,0xff,
+    0xe3,0xdd,0xfd,0xfd,0xcd,0xdd,0xc3,0xff,
     0xdd,0xdd,0xdd,0xc1,0xdd,0xdd,0xdd,0xff,
     0xe3,0xf7,0xf7,0xf7,0xf7,0xf7,0xe3,0xff,
-    0xdf,0xdf,0xdf,0xdf,0xdf,0xdd,0xe3,0xff,
-    0xdd,0xed,0xf5,0xf9,0xf5,0xed,0xdd,0xff,
+    0xc7,0xef,0xef,0xef,0xef,0xed,0xf3,0xff,
+    0xdd,0xed,0xf5,0xe9,0xed,0xdd,0xdd,0xff,
     0xfd,0xfd,0xfd,0xfd,0xfd,0xfd,0xc1,0xff,
     0xdd,0xc9,0xd5,0xd5,0xdd,0xdd,0xdd,0xff,
     0xdd,0xdd,0xd9,0xd5,0xcd,0xdd,0xdd,0xff,
@@ -217,7 +215,7 @@ const apple2e_charset = new Uint8Array([
     0xe3,0xdd,0xfd,0xe3,0xdf,0xdd,0xe3,0xff,
     0xc1,0xf7,0xf7,0xf7,0xf7,0xf7,0xf7,0xff,
     0xdd,0xdd,0xdd,0xdd,0xdd,0xdd,0xe3,0xff,
-    0xdd,0xdd,0xdd,0xdd,0xdd,0xeb,0xf7,0xff,
+    0xdd,0xdd,0xdd,0xdd,0xeb,0xeb,0xf7,0xff,
     0xdd,0xdd,0xdd,0xd5,0xd5,0xc9,0xdd,0xff,
     0xdd,0xdd,0xeb,0xf7,0xeb,0xdd,0xdd,0xff,
     0xdd,0xdd,0xeb,0xf7,0xf7,0xf7,0xf7,0xff,
@@ -244,13 +242,13 @@ const apple2e_charset = new Uint8Array([
     0xff,0xff,0xe1,0xdd,0xdd,0xdd,0xdd,0xff,
     0xff,0xff,0xe3,0xdd,0xdd,0xdd,0xe3,0xff,
     0xff,0xff,0xe1,0xdd,0xdd,0xe1,0xfd,0xfd,
-    0xff,0xff,0xc3,0xdd,0xdd,0xc3,0xdf,0xdf,
-    0xff,0xff,0xc5,0xf9,0xfd,0xfd,0xfd,0xff,
+    0x7f,0x7f,0x43,0x5d,0x5d,0x43,0x5f,0x1f,
+    0x7f,0x7f,0x65,0x59,0x7d,0x7d,0x7d,0x7f,
     0xff,0xff,0xc3,0xfd,0xe3,0xdf,0xe1,0xff,
     0xfb,0xfb,0xe1,0xfb,0xfb,0xdb,0xe7,0xff,
     0xff,0xff,0xdd,0xdd,0xdd,0xcd,0xd3,0xff,
     0xff,0xff,0xdd,0xdd,0xdd,0xeb,0xf7,0xff,
-    0xff,0xff,0xdd,0xdd,0xd5,0xd5,0xc9,0xff,
+    0xff,0xff,0x5d,0x5d,0x55,0x55,0x6b,0x7f,
     0xff,0xff,0xdd,0xeb,0xf7,0xeb,0xdd,0xff,
     0xff,0xff,0xdd,0xdd,0xdd,0xc3,0xdf,0xe3,
     0xff,0xff,0xc1,0xef,0xf7,0xfb,0xc1,0xff,
@@ -258,7 +256,7 @@ const apple2e_charset = new Uint8Array([
     0xf7,0xf7,0xf7,0xf7,0xf7,0xf7,0xf7,0xf7,
     0xf1,0xe7,0xe7,0xcf,0xe7,0xe7,0xf1,0xff,
     0xd3,0xe5,0xff,0xff,0xff,0xff,0xff,0xff,
-    0xff,0xd5,0xeb,0xd5,0xeb,0xd5,0xff,0xff,
+    0x6f,0x77,0x49,0x3e,0x5e,0x5e,0x36,0x49,
     0xff,0xff,0xff,0xff,0xff,0xff,0xff,0xff,
     0xfe,0xfe,0xee,0xbb,0xfe,0xfe,0xff,0xff,
     0xfd,0xfd,0xdd,0x77,0xfd,0xfd,0xff,0xff,
@@ -517,4 +515,4 @@ const apple2e_charset = new Uint8Array([
     0x80,0x80,0x00,0x00,0x80,0x80,0x00,0x00
 ]) as ReadonlyUint8Array;
 
-export default apple2e_charset;
+export default rmfont_charset;
