+++ js/roms/system/apple2j.ts
@@ -1,519 +1,7 @@
-import { ReadonlyUint8Array } from '../types';
-import ROM from './rom';
+import { ReadonlyUint8Array } from '../../types';
+import ROM from '../rom';
 
 const rom = new Uint8Array([
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0xc0,0x06,0xd0,0x03,0x4c,0x88,0xc2,0xc0,
-    0x07,0xd0,0x03,0x4c,0x6e,0xc2,0x98,0x48,
-    0x20,0x24,0xcb,0xd0,0x0a,0xad,0xfb,0x04,
-    0x29,0x28,0xd0,0x03,0x4c,0xa4,0xc1,0x68,
-    0xa8,0xa9,0xc1,0x48,0xb9,0xea,0xcf,0x48,
-    0x60,0xa4,0x24,0xa5,0x25,0x48,0x20,0x24,
-    0xfc,0x20,0xf4,0xc2,0xa0,0x00,0x68,0x69,
-    0x00,0xc5,0x23,0x90,0xf0,0x20,0x22,0xfc,
-    0x4c,0xeb,0xc2,0xa5,0x22,0x85,0x25,0xa0,
-    0x00,0x84,0x24,0xf0,0xe0,0xa5,0x22,0x48,
-    0x20,0x24,0xfc,0xa5,0x28,0x85,0x2a,0xa5,
-    0x29,0x85,0x2b,0xa4,0x21,0x88,0x68,0x69,
-    0x01,0xc5,0x23,0xb0,0x0d,0x48,0x20,0x24,
-    0xfc,0xb1,0x28,0x91,0x2a,0x88,0x10,0xf9,
-    0x30,0xe1,0xa0,0x00,0x20,0xf4,0xc2,0x20,
-    0x22,0xfc,0x4c,0xeb,0xc2,0xa4,0x24,0xa9,
-    0xa0,0x91,0x28,0xc8,0xc4,0x21,0x90,0xf9,
-    0xb0,0x17,0xa9,0x28,0x85,0x21,0xa9,0x18,
-    0x85,0x23,0xa9,0x17,0x85,0x25,0x20,0x22,
-    0xfc,0x4c,0xeb,0xc2,0xa4,0x1f,0x20,0xf4,
-    0xc2,0x4c,0xeb,0xc2,0x68,0xa8,0xad,0xfb,
-    0x04,0x29,0xfe,0x8d,0xfb,0x04,0x68,0x8d,
-    0x78,0x04,0x68,0x48,0x4a,0x4a,0x4a,0xad,
-    0x78,0x04,0x48,0xb0,0x08,0xad,0xfb,0x04,
-    0x09,0x01,0x8d,0xfb,0x04,0xa5,0x25,0x8d,
-    0xfb,0x05,0x4c,0xff,0xc1,0x20,0xa4,0xcc,
-    0x4c,0xeb,0xc2,0x20,0x48,0xcd,0x4c,0xeb,
-    0xc2,0xa4,0x1f,0x20,0x4e,0xcd,0x4c,0xeb,
-    0xc2,0x20,0x23,0xcd,0x4c,0xeb,0xc2,0x4c,
-    0x19,0xc2,0x4c,0x34,0xc2,0x20,0x42,0xcd,
-    0xad,0x7b,0x05,0x85,0x24,0x8d,0x7b,0x04,
-    0xad,0xfb,0x05,0x85,0x25,0x10,0x2f,0x20,
-    0x51,0xcb,0xa5,0x24,0x2c,0x1f,0xc0,0x10,
-    0x05,0xcd,0x7b,0x04,0xf0,0x03,0x8d,0x7b,
-    0x05,0xa9,0xc1,0x48,0xb9,0xf3,0xcf,0x48,
-    0x60,0xa9,0x50,0x2c,0x1f,0xc0,0x30,0x01,
-    0x4a,0x85,0x21,0xa9,0x18,0x85,0x23,0xa9,
-    0x17,0x8d,0xfb,0x05,0x85,0x25,0x20,0x51,
-    0xcb,0x4c,0xeb,0xc2,0xa9,0xff,0x8d,0xfb,
-    0x04,0xad,0x5d,0xc0,0xad,0x5f,0xc0,0xad,
-    0x62,0xc0,0x30,0x1d,0xad,0x61,0xc0,0x10,
-    0x1b,0xa0,0xb0,0xa9,0x00,0x85,0x3c,0xa9,
-    0xbf,0x38,0x85,0x3d,0x91,0x3c,0x88,0x91,
-    0x3c,0xe9,0x01,0xc9,0x01,0xd0,0xf3,0xf0,
-    0x03,0x4c,0x01,0xc4,0x20,0x24,0xcb,0xf0,
-    0x14,0x8d,0x0b,0xc0,0xd0,0x0f,0x29,0xdf,
-    0xa0,0x03,0xd9,0x80,0xc2,0xd0,0x03,0xb9,
-    0x84,0xc2,0x88,0x10,0xf5,0x4c,0xeb,0xc2,
-    0x88,0x95,0x8a,0x8b,0xca,0xcb,0xcd,0xc9,
-    0x8d,0x78,0x04,0x68,0xa8,0x68,0x48,0x6a,
-    0x6a,0x6a,0x98,0x48,0x8a,0x48,0xb8,0xb0,
-    0x03,0x2c,0x00,0xcf,0xa9,0xff,0xa4,0x24,
-    0x91,0x28,0x20,0xc6,0xc2,0xb0,0x0e,0xad,
-    0x78,0x04,0xa4,0x24,0x91,0x28,0x20,0xc6,
-    0xc2,0xb0,0x02,0x90,0xe7,0xad,0x78,0x04,
-    0xa4,0x24,0x91,0x28,0x68,0xaa,0xad,0x00,
-    0xc0,0x8d,0x10,0xc0,0x30,0x25,0xa2,0x0c,
-    0x70,0x02,0xa2,0x31,0xa0,0x00,0x50,0x05,
-    0x08,0x20,0x75,0xfc,0x28,0xe6,0x4e,0xd0,
-    0x02,0xe6,0x4f,0xad,0x00,0xc0,0x30,0x09,
-    0x88,0xd0,0xeb,0xca,0xd0,0xe6,0x18,0x90,
-    0x01,0x38,0x60,0x28,0x30,0x03,0x4c,0x29,
-    0xfd,0x4c,0x2c,0xfd,0xa9,0xa0,0x91,0x28,
-    0xc8,0xc4,0x21,0x90,0xf9,0x60,0x00,0x00,
-    0x2c,0x58,0xff,0x70,0x12,0x38,0x90,0x18,
-    0xb8,0x50,0x0c,0x01,0x88,0x4b,0x51,0x57,
-    0x5d,0x4c,0x63,0xc3,0x4c,0xb0,0xc3,0x8d,
-    0x7b,0x06,0x48,0x98,0x48,0x8a,0x48,0x08,
-    0xad,0xfb,0x04,0x29,0xfe,0x8d,0xfb,0x04,
-    0x68,0x48,0x29,0x04,0xd0,0x08,0xad,0xfb,
-    0x04,0x09,0x01,0x8d,0xfb,0x04,0xad,0xff,
-    0xcf,0xa5,0x25,0x8d,0xfb,0x05,0x20,0xeb,
-    0xc3,0x28,0x08,0x70,0x03,0x4c,0x66,0xc8,
-    0x4c,0x03,0xc8,0x20,0xeb,0xc3,0x4c,0x4f,
-    0xca,0x20,0xeb,0xc3,0x4c,0x74,0xca,0x20,
-    0xeb,0xc3,0x4c,0x8e,0xca,0x20,0xeb,0xc3,
-    0x4c,0x94,0xc9,0x48,0x98,0x48,0xad,0x13,
-    0xc0,0x48,0xad,0x14,0xc0,0x48,0x90,0x08,
-    0x8d,0x02,0xc0,0x8d,0x05,0xc0,0xb0,0x06,
-    0x8d,0x04,0xc0,0x8d,0x03,0xc0,0xa0,0x00,
-    0xb1,0x3c,0x91,0x42,0xe6,0x42,0xd0,0x02,
-    0xe6,0x43,0xa5,0x3c,0xc5,0x3e,0xa5,0x3d,
-    0xe5,0x3f,0xe6,0x3c,0xd0,0x02,0xe6,0x3d,
-    0x90,0xe6,0x8d,0x04,0xc0,0x68,0x10,0x03,
-    0x8d,0x05,0xc0,0x8d,0x02,0xc0,0x68,0x10,
-    0x03,0x8d,0x03,0xc0,0x68,0xa8,0x68,0x60,
-    0x48,0xad,0xed,0x03,0x48,0xad,0xee,0x03,
-    0x48,0x90,0x0a,0x8d,0x03,0xc0,0x8d,0x05,
-    0xc0,0x50,0x19,0x70,0x08,0x8d,0x02,0xc0,
-    0x8d,0x04,0xc0,0x50,0x0f,0x68,0x8d,0xee,
-    0x03,0x68,0x8d,0xed,0x03,0x68,0x8d,0x09,
-    0xc0,0x6c,0xed,0x03,0x68,0x8d,0xee,0x03,
-    0x68,0x8d,0xed,0x03,0x68,0x8d,0x08,0xc0,
-    0x6c,0xed,0x03,0x48,0xa9,0xc3,0x8d,0xf8,
-    0x07,0x68,0x60,0x00,0x00,0x00,0x00,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x1b,0xad,0x61,0xc0,0x85,0x05,0xa2,0xff,
-    0x9a,0xa9,0xf0,0x85,0x36,0xa9,0xfd,0x85,
-    0x37,0xad,0x82,0xc0,0x20,0x58,0xfc,0xa9,
-    0x00,0x85,0x09,0x85,0x0a,0x85,0x08,0xaa,
-    0xbd,0x11,0xc0,0x5d,0xee,0xc7,0x10,0x03,
-    0x20,0x13,0xc5,0xe8,0xe0,0x08,0xd0,0xf0,
-    0x8d,0x8b,0xc0,0x2c,0x11,0xc0,0x8d,0x83,
-    0xc0,0x10,0x03,0x20,0x13,0xc5,0xe8,0x8d,
-    0x80,0xc0,0x2c,0x12,0xc0,0x8d,0x82,0xc0,
-    0x30,0x03,0x20,0x13,0xc5,0xe8,0x8d,0x03,
-    0xc0,0x2c,0x13,0xc0,0x8d,0x02,0xc0,0x30,
-    0x03,0x20,0x13,0xc5,0xe8,0x8d,0x05,0xc0,
-    0x2c,0x14,0xc0,0x8d,0x04,0xc0,0x30,0x03,
-    0x20,0x13,0xc5,0xe8,0x8d,0x09,0xc0,0x2c,
-    0x16,0xc0,0x8d,0x08,0xc0,0x30,0x03,0x20,
-    0x13,0xc5,0xe8,0x8d,0x0b,0xc0,0x2c,0x17,
-    0xc0,0x8d,0x0a,0xc0,0x30,0x03,0x20,0x13,
-    0xc5,0xe8,0x8d,0x01,0xc0,0x2c,0x18,0xc0,
-    0x8d,0x00,0xc0,0x30,0x03,0x20,0x13,0xc5,
-    0x20,0x8e,0xfd,0xa2,0x00,0xa9,0xff,0x85,
-    0x08,0xbd,0x1a,0xc0,0x5d,0xf6,0xc7,0x10,
-    0x03,0x20,0x13,0xc5,0xe8,0xe0,0x06,0xd0,
-    0xf0,0x8d,0x0d,0xc0,0x2c,0x1f,0xc0,0x8d,
-    0x0c,0xc0,0x30,0x03,0x20,0x13,0xc5,0xe8,
-    0x8d,0x0f,0xc0,0x2c,0x1e,0xc0,0x8d,0x0e,
-    0xc0,0x30,0x03,0x20,0x13,0xc5,0xe8,0x8d,
-    0x50,0xc0,0x2c,0x1a,0xc0,0x8d,0x51,0xc0,
-    0x10,0x03,0x20,0x13,0xc5,0xe8,0x8d,0x55,
-    0xc0,0x2c,0x1c,0xc0,0x8d,0x54,0xc0,0x30,
-    0x03,0x20,0x13,0xc5,0xe8,0x8d,0x53,0xc0,
-    0x2c,0x1b,0xc0,0x8d,0x52,0xc0,0x30,0x03,
-    0x20,0x13,0xc5,0xe8,0x8d,0x57,0xc0,0x2c,
-    0x1d,0xc0,0x8d,0x56,0xc0,0x30,0x03,0x20,
-    0x13,0xc5,0xa5,0x09,0x05,0x0a,0xf0,0x2f,
-    0x4c,0x10,0xc5,0x86,0x04,0xa5,0x08,0xd0,
-    0x0d,0xa5,0x09,0xd0,0x14,0xa2,0x44,0x20,
-    0x40,0xc7,0xe6,0x09,0xd0,0x0b,0xa5,0x0a,
-    0xd0,0x07,0xa2,0x51,0x20,0x40,0xc7,0xe6,
-    0x0a,0xa5,0x04,0x20,0xe3,0xfd,0xa9,0xa0,
-    0x20,0xf0,0xfd,0xa5,0x04,0xaa,0x60,0xa9,
-    0xc1,0x85,0x01,0xa9,0x00,0x85,0x00,0xa8,
-    0x18,0x71,0x00,0xc8,0xd0,0xfa,0xe6,0x01,
-    0xa6,0x01,0xe0,0xc4,0xd0,0x03,0xc8,0xd0,
-    0xef,0xe0,0xe0,0xf0,0x02,0xd0,0xe9,0xcd,
-    0x00,0xc4,0xf0,0x08,0xa2,0x00,0x20,0x40,
-    0xc7,0x4c,0x69,0xc5,0xa0,0x00,0x98,0x18,
-    0x71,0x00,0xc8,0xf0,0x0b,0xc0,0xff,0xd0,
-    0xf6,0xa6,0x01,0xe0,0xf7,0xd0,0xf0,0xc8,
-    0xe6,0x01,0xd0,0xeb,0xcd,0xff,0xf7,0xf0,
-    0x08,0xa2,0x08,0x20,0x40,0xc7,0x4c,0x8e,
-    0xc5,0xad,0x50,0xc0,0xad,0x57,0xc0,0xad,
-    0x52,0xc0,0xad,0x83,0xc0,0xad,0x83,0xc0,
-    0xa9,0x01,0x85,0x04,0xa9,0xff,0x85,0x03,
-    0xa9,0x00,0xaa,0x85,0x00,0x85,0x01,0xa0,
-    0x06,0x91,0x00,0xc8,0xd0,0xfb,0x24,0x05,
-    0x10,0x03,0x8d,0x30,0xc0,0xe6,0x01,0xf0,
-    0x2a,0xa5,0x01,0xc9,0xc0,0xd0,0x06,0xa9,
-    0xd0,0x85,0x01,0x30,0x1b,0xc9,0xe0,0xd0,
-    0x17,0x2c,0x11,0xc0,0x10,0x0c,0xad,0x8b,
-    0xc0,0xad,0x8b,0xc0,0xa9,0xd0,0x85,0x01,
-    0x30,0x06,0xad,0x83,0xc0,0xad,0x83,0xc0,
-    0x8a,0xf0,0xc6,0x85,0x02,0x49,0xff,0xaa,
-    0xa9,0x00,0x85,0x01,0xa0,0x06,0x24,0x05,
-    0x10,0x03,0x8d,0x30,0xc0,0xc6,0x04,0xb1,
-    0x00,0xc5,0x02,0xf0,0x03,0x4c,0xaf,0xc6,
-    0xe6,0x04,0x8a,0x91,0x00,0xd1,0x00,0xf0,
-    0x03,0x4c,0xaf,0xc6,0xc8,0xd0,0xdf,0xe6,
-    0x01,0xf0,0x2a,0xa5,0x01,0xc9,0xc0,0xd0,
-    0x06,0xa9,0xd0,0x85,0x01,0x30,0xcf,0xc9,
-    0xe0,0xd0,0xcb,0x2c,0x11,0xc0,0x10,0x0c,
-    0xad,0x8b,0xc0,0xad,0x8b,0xc0,0xa9,0xd0,
-    0x85,0x01,0x30,0xba,0xad,0x83,0xc0,0xad,
-    0x83,0xc0,0x4c,0xf6,0xc5,0x85,0x02,0x49,
-    0xff,0xaa,0xf0,0xa4,0xa9,0xff,0x85,0x01,
-    0xa8,0x24,0x05,0x10,0x03,0x8d,0x30,0xc0,
-    0xc6,0x04,0xb1,0x00,0xc5,0x02,0xd0,0x4f,
-    0xe6,0x04,0x8a,0x91,0x00,0xd1,0x00,0xd0,
-    0x46,0x88,0xc4,0x03,0xd0,0xe3,0xc6,0x01,
-    0xa5,0x01,0xc9,0xff,0xf0,0x2d,0xc9,0xcf,
-    0xd0,0x1d,0x2c,0x11,0xc0,0x10,0x0c,0xad,
-    0x8b,0xc0,0xad,0x8b,0xc0,0xa9,0xdf,0x85,
-    0x01,0x30,0xc6,0xad,0x83,0xc0,0xad,0x83,
-    0xc0,0xa9,0xbf,0x85,0x01,0x30,0xba,0xc9,
-    0x00,0xf0,0x02,0xd0,0xb4,0xa9,0x06,0x85,
-    0x03,0xd0,0xae,0x8a,0xf0,0x71,0x85,0x02,
-    0x85,0x03,0x49,0xff,0xaa,0xf0,0x9d,0xad,
-    0x82,0xc0,0x20,0xe3,0xc7,0xad,0x51,0xc0,
-    0xa2,0x11,0x20,0x40,0xc7,0xa5,0x04,0xf0,
-    0x0b,0xa5,0x02,0x49,0xff,0x51,0x00,0x85,
-    0x06,0x4c,0xd2,0xc6,0xa5,0x02,0x51,0x00,
-    0x85,0x06,0xa5,0x06,0x0a,0x90,0x05,0xa2,
-    0x20,0x20,0x40,0xc7,0x0a,0x90,0x05,0xa2,
-    0x25,0x20,0x40,0xc7,0x0a,0x90,0x05,0xa2,
-    0x2a,0x20,0x40,0xc7,0x0a,0x90,0x05,0xa2,
-    0x2f,0x20,0x40,0xc7,0x0a,0x90,0x05,0xa2,
-    0x34,0x20,0x40,0xc7,0x0a,0x90,0x05,0xa2,
-    0x38,0x20,0x40,0xc7,0x0a,0x90,0x05,0xa2,
-    0x3c,0x20,0x40,0xc7,0x0a,0x90,0x05,0xa2,
-    0x40,0x20,0x40,0xc7,0x4c,0x14,0xc7,0xad,
-    0x82,0xc0,0x20,0xe3,0xc7,0xad,0x51,0xc0,
-    0xad,0x61,0xc0,0x2d,0x62,0xc0,0x30,0x08,
-    0xa2,0x16,0x20,0x40,0xc7,0x4c,0x2d,0xc7,
-    0xa2,0x00,0xbd,0xae,0xc7,0x9d,0x00,0x01,
-    0xe8,0xe0,0x35,0xd0,0xf5,0x4c,0x00,0x01,
-    0x85,0x07,0xbd,0x50,0xc7,0xf0,0x06,0x20,
-    0xf0,0xfd,0xe8,0xd0,0xf5,0xa5,0x07,0x60,
-    0xd2,0xcf,0xcd,0xba,0xc5,0xb8,0x8d,0x00,
-    0xd2,0xcf,0xcd,0xba,0xc5,0xb1,0xb0,0x8d,
-    0x00,0xd2,0xc1,0xcd,0xba,0x00,0xcb,0xc5,
-    0xd2,0xce,0xc5,0xcc,0xa0,0xcf,0xcb,0x00,
-    0xc6,0xb1,0xb3,0xa0,0x00,0xc6,0xb1,0xb2,
-    0xa0,0x00,0xc6,0xb1,0xb1,0xa0,0x00,0xc6,
-    0xb1,0xb0,0xa0,0x00,0xc6,0xb9,0xa0,0x00,
-    0xc6,0xb8,0xa0,0x00,0xc6,0xb7,0xa0,0x00,
-    0xc6,0xb6,0xa0,0x00,0xcd,0xcd,0xd5,0xa0,
-    0xc6,0xcc,0xc1,0xc7,0xa0,0xc5,0xb4,0xba,
-    0x00,0xc9,0xcf,0xd5,0xa0,0xc6,0xcc,0xc1,
-    0xc7,0xa0,0xc5,0xb5,0xba,0x00,0x8d,0x06,
-    0xc0,0x8d,0xff,0xcf,0xa0,0x00,0xa9,0xc0,
-    0x85,0x01,0xa9,0x00,0x85,0x00,0xa0,0x90,
-    0xb1,0x00,0x91,0x00,0xc8,0xd0,0xf9,0xe6,
-    0x01,0xa5,0x01,0xc9,0xd0,0xd0,0xf1,0xa9,
-    0xc0,0x85,0x01,0xa0,0x79,0xb1,0x00,0x88,
-    0xd0,0xfb,0x8d,0x07,0xc0,0x8d,0x51,0xc0,
-    0x4c,0x01,0xc4,0x20,0x2f,0xfb,0xa9,0xff,
-    0x85,0x32,0x20,0x58,0xfc,0x60,0x80,0x00,
-    0x00,0x00,0x80,0x00,0x00,0x00,0x80,0x00,
-    0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00,
-    0x4c,0x4a,0xca,0xa9,0x06,0xcd,0xb3,0xfb,
-    0xf0,0x0c,0x20,0x78,0xcf,0xcd,0xb3,0xfb,
-    0xf0,0x04,0x78,0x4c,0x13,0xc8,0xa9,0xc3,
-    0x85,0x37,0x85,0x39,0xa9,0x05,0x85,0x38,
-    0xa9,0x07,0x85,0x36,0xa9,0x00,0x85,0x20,
-    0xa9,0x00,0x2c,0x1a,0xc0,0x30,0x02,0xa9,
-    0x14,0x85,0x22,0xa9,0x18,0x85,0x23,0xa9,
-    0x28,0x85,0x21,0xa5,0x24,0x8d,0x7b,0x04,
-    0xa9,0x01,0x2d,0xfb,0x04,0x8d,0xfb,0x04,
-    0x4c,0x50,0xc8,0x00,0x00,0x4c,0x51,0xc3,
-    0x20,0x24,0xcb,0xd0,0x08,0x06,0x21,0x8d,
-    0x01,0xc0,0x8d,0x0d,0xc0,0x8d,0x0f,0xc0,
-    0x20,0x42,0xcd,0x28,0x18,0x08,0x2c,0x1f,
-    0xc0,0x10,0x09,0x8d,0x01,0xc0,0xa5,0x21,
-    0x29,0xfe,0x85,0x21,0xa5,0x24,0xcd,0x7b,
-    0x04,0xf0,0x03,0x8d,0x7b,0x05,0xa9,0x06,
-    0xcd,0xb3,0xfb,0xf0,0x0b,0x20,0x78,0xcf,
-    0xcd,0xb3,0xfb,0xf0,0x03,0x4c,0x13,0xc8,
-    0x28,0x90,0x03,0x4c,0xf6,0xc8,0xad,0xfb,
-    0x04,0x29,0xbf,0x8d,0xfb,0x04,0x4c,0xa1,
-    0xc8,0xad,0x7b,0x06,0xc9,0x8d,0xd0,0x18,
-    0xac,0x00,0xc0,0x10,0x13,0xc0,0x93,0xd0,
-    0x0f,0x2c,0x10,0xc0,0xac,0x00,0xc0,0x10,
-    0xfb,0xc0,0x83,0xf0,0x03,0x2c,0x10,0xc0,
-    0x29,0x7f,0xc9,0x20,0xb0,0x06,0x20,0x99,
-    0xcb,0x4c,0xe2,0xc8,0xac,0x7b,0x05,0xad,
-    0x7b,0x06,0x20,0xf2,0xce,0xee,0x7b,0x05,
-    0xad,0x7b,0x05,0xc5,0x21,0x90,0x03,0x20,
-    0xec,0xcb,0xad,0x7b,0x05,0x20,0xaf,0xce,
-    0xad,0xfb,0x05,0x85,0x25,0x68,0xaa,0x68,
-    0xa8,0x68,0xad,0x7b,0x06,0x60,0xad,0xfb,
-    0x04,0x09,0x40,0x8d,0xfb,0x04,0xad,0x7b,
-    0x06,0xa4,0x24,0x91,0x28,0x20,0xdd,0xce,
-    0x20,0x15,0xcb,0x8d,0x7b,0x06,0x20,0xdd,
-    0xce,0xc9,0x9b,0xf0,0x03,0x4c,0xb7,0xc9,
-    0x20,0x52,0xcf,0x20,0x15,0xcb,0x20,0x65,
-    0xcf,0x29,0x7f,0xc9,0x60,0x90,0x02,0x29,
-    0xdf,0xa0,0x11,0xd9,0x72,0xc9,0xf0,0x05,
-    0x88,0x10,0xf8,0x30,0x10,0xb9,0x83,0xc9,
-    0x29,0x7f,0x20,0x99,0xcb,0xb9,0x83,0xc9,
-    0x30,0xd6,0x4c,0x05,0xc9,0xc9,0x11,0xd0,
-    0x0b,0x20,0xaa,0xcd,0xa9,0x98,0x8d,0x7b,
-    0x06,0x4c,0xe2,0xc8,0xc9,0x52,0xd0,0x0b,
-    0xad,0xfb,0x04,0x09,0x80,0x8d,0xfb,0x04,
-    0x4c,0x05,0xc9,0xc9,0x54,0xd0,0xf9,0xad,
-    0xfb,0x04,0x29,0x7f,0x8d,0xfb,0x04,0x4c,
-    0x05,0xc9,0x40,0x41,0x42,0x43,0x44,0x45,
-    0x46,0x49,0x4a,0x4b,0x4d,0x34,0x38,0x08,
-    0x0a,0x0b,0x15,0x0c,0x1c,0x08,0x0a,0x1f,
-    0x1d,0x0b,0x9f,0x88,0x9c,0x8a,0x11,0x12,
-    0x88,0x8a,0x9f,0x9c,0xaa,0x20,0xc8,0xcf,
-    0x8a,0xd0,0x03,0x38,0xb0,0x16,0xc9,0x01,
-    0xf0,0x0e,0xa2,0x03,0x18,0x60,0x00,0x00,
-    0x00,0x00,0xad,0x7b,0x06,0x4c,0x57,0xc3,
-    0xad,0x00,0xc0,0x0a,0xa2,0x00,0x60,0xc9,
-    0x95,0xd0,0x0b,0xac,0x7b,0x05,0x20,0x01,
-    0xcf,0x09,0x80,0x8d,0x7b,0x06,0xad,0xfb,
-    0x04,0x29,0x10,0xd0,0x12,0xad,0x7b,0x06,
-    0xc9,0xa2,0xf0,0x23,0xc9,0x88,0xd0,0x32,
-    0x20,0x27,0xca,0xd0,0x2d,0xf0,0x18,0xad,
-    0x7b,0x06,0xc9,0xa2,0xf0,0x1c,0xc9,0x98,
-    0xf0,0x18,0xc9,0x8d,0xf0,0x14,0xc9,0x88,
-    0xd0,0x18,0x20,0x27,0xca,0xd0,0x13,0xad,
-    0xfb,0x04,0x49,0x10,0x8d,0xfb,0x04,0x4c,
-    0x0a,0xca,0xad,0xfb,0x04,0x29,0xef,0x8d,
-    0xfb,0x04,0xad,0xfb,0x04,0x29,0x80,0xf0,
-    0x13,0xad,0xfb,0x04,0x29,0x10,0xd0,0x0c,
-    0xad,0x7b,0x06,0xc9,0xe0,0x90,0x05,0x29,
-    0xdf,0x8d,0x7b,0x06,0x4c,0xe2,0xc8,0xad,
-    0xfb,0x05,0x0d,0x7b,0x05,0xf0,0x1a,0x98,
-    0x48,0x20,0xdb,0xcb,0xac,0x7b,0x05,0x20,
-    0x01,0xcf,0x09,0x80,0x8d,0x78,0x04,0x20,
-    0x26,0xcc,0x68,0xa8,0xad,0x78,0x04,0xc9,
-    0xa2,0x60,0xa9,0x22,0x4c,0x51,0xca,0xa9,
-    0x20,0x8d,0xfb,0x04,0x20,0x9b,0xcd,0x20,
-    0xc8,0xcf,0x20,0x24,0xcb,0xf0,0x03,0xa2,
-    0x09,0x60,0x8d,0x01,0xc0,0x8d,0x0d,0xc0,
-    0x8d,0x0f,0xc0,0x20,0x42,0xcd,0x20,0xdd,
-    0xce,0xa2,0x00,0x60,0x20,0xc8,0xcf,0x20,
-    0x15,0xcb,0x29,0x7f,0x8d,0x7b,0x06,0xa2,
-    0x00,0xad,0xfb,0x04,0x29,0x02,0xf0,0x02,
-    0xa2,0xc3,0xad,0x7b,0x06,0x60,0x8d,0x7b,
-    0x06,0x20,0xc8,0xcf,0x20,0xdd,0xce,0xad,
-    0xfb,0x04,0x29,0x08,0xf0,0x2d,0xad,0xfb,
-    0x06,0x10,0x0c,0xad,0x7b,0x06,0x38,0xe9,
-    0x20,0x8d,0xfb,0x06,0x4c,0x0f,0xcb,0xad,
-    0x7b,0x06,0x38,0xe9,0x20,0x8d,0xfb,0x05,
-    0x20,0x51,0xcb,0xad,0xfb,0x06,0x8d,0x7b,
-    0x05,0xad,0xfb,0x04,0x29,0xf7,0x8d,0xfb,
-    0x04,0xd0,0x44,0xad,0x7b,0x06,0xc9,0x1e,
-    0xf0,0x0a,0xc9,0x20,0xb0,0x15,0x20,0x99,
-    0xcb,0x4c,0x0f,0xcb,0xad,0xfb,0x04,0x09,
-    0x08,0x8d,0xfb,0x04,0xa9,0xff,0x8d,0xfb,
-    0x06,0x30,0x24,0x09,0x80,0xac,0x7b,0x05,
-    0x20,0xf2,0xce,0xee,0x7b,0x05,0xad,0x7b,
-    0x05,0xc5,0x21,0x90,0x12,0xad,0xfb,0x04,
-    0x29,0x01,0xf0,0x05,0xce,0x7b,0x05,0xd0,
-    0x06,0x20,0xec,0xcb,0x20,0x91,0xcc,0x20,
-    0xdd,0xce,0xa2,0x00,0x60,0xe6,0x4e,0xd0,
-    0x02,0xe6,0x4f,0xad,0x00,0xc0,0x10,0xf5,
-    0x8d,0x10,0xc0,0x60,0xad,0x1c,0xc0,0x0a,
-    0xa9,0x88,0x2c,0x18,0xc0,0x8d,0x01,0xc0,
-    0x08,0x78,0x08,0x8d,0x55,0xc0,0xac,0x00,
-    0x04,0x8d,0x00,0x04,0xad,0x00,0x04,0x8c,
-    0x00,0x04,0x28,0xb0,0x03,0x8d,0x54,0xc0,
-    0x30,0x03,0x8d,0x00,0xc0,0x28,0xc9,0x88,
-    0x60,0x18,0x90,0x01,0x38,0x48,0xb0,0x03,
-    0xad,0xfb,0x05,0x48,0x4a,0x29,0x03,0x09,
-    0x04,0x85,0x29,0x8d,0xfb,0x07,0x68,0x29,
-    0x18,0x90,0x02,0x69,0x7f,0x85,0x28,0x0a,
-    0x0a,0x05,0x28,0x85,0x28,0xa5,0x20,0x08,
-    0x2c,0x1f,0xc0,0x10,0x01,0x4a,0x28,0x65,
-    0x28,0x85,0x28,0x8d,0x7b,0x07,0xad,0xfb,
-    0x04,0x29,0x01,0xf0,0x0a,0xad,0xfb,0x04,
-    0x29,0x20,0xd0,0x03,0x20,0x75,0xfc,0x68,
-    0x60,0x8d,0x78,0x04,0x48,0x98,0x48,0xac,
-    0x78,0x04,0xc0,0x07,0x90,0x05,0xb9,0x71,
-    0xcc,0xd0,0x03,0x38,0xb0,0x04,0x20,0xb6,
-    0xcb,0x18,0x68,0xa8,0x68,0x60,0x48,0xb9,
-    0x58,0xcc,0x48,0x60,0xa9,0x40,0x20,0xcf,
-    0xcb,0xa0,0xc0,0xa9,0x0c,0x20,0xcf,0xcb,
-    0xad,0x30,0xc0,0x88,0xd0,0xf5,0x60,0x38,
-    0x48,0xe9,0x01,0xd0,0xfc,0x68,0xe9,0x01,
-    0xd0,0xf6,0x60,0xce,0x7b,0x05,0x10,0x0b,
-    0xa5,0x21,0x8d,0x7b,0x05,0xce,0x7b,0x05,
-    0x20,0x34,0xcc,0x60,0xad,0xfb,0x04,0x29,
-    0x20,0xd0,0x0a,0xad,0xfb,0x04,0x29,0x40,
-    0xf0,0x03,0x20,0x48,0xcd,0xa9,0x00,0x8d,
-    0x7b,0x05,0xad,0xfb,0x04,0x29,0x20,0xd0,
-    0x03,0x20,0x91,0xcc,0x60,0xa5,0x22,0x8d,
-    0xfb,0x05,0xa9,0x00,0x8d,0x7b,0x05,0x4c,
-    0x51,0xcb,0xa4,0x21,0x88,0xa9,0xa0,0x20,
-    0xf2,0xce,0x88,0x10,0xfa,0x60,0xee,0x7b,
-    0x05,0xad,0x7b,0x05,0xc5,0x21,0x90,0x03,
-    0x20,0xec,0xcb,0x60,0xce,0xfb,0x05,0x30,
-    0x07,0xad,0xfb,0x05,0xc5,0x22,0xb0,0x05,
-    0xee,0xfb,0x05,0xf0,0x03,0x20,0x51,0xcb,
-    0x60,0xad,0xfb,0x04,0x29,0xfb,0xa0,0xff,
-    0xd0,0x07,0xad,0xfb,0x04,0x09,0x04,0xa0,
-    0x7f,0x8d,0xfb,0x04,0x84,0x32,0x60,0xbb,
-    0xda,0x00,0x90,0x22,0x41,0xeb,0x48,0x51,
-    0x00,0x58,0x76,0x00,0x00,0x8f,0xa9,0xa3,
-    0x00,0x0c,0x19,0x00,0x25,0x47,0x00,0x33,
-    0xcb,0xcb,0x00,0xcc,0xcd,0xcd,0xcb,0xcc,
-    0xcc,0x00,0xcd,0xcd,0x00,0x00,0xcd,0xcc,
-    0xcc,0x00,0xcc,0xcc,0x00,0xcc,0xcd,0x00,
-    0xcc,0xee,0xfb,0x05,0xad,0xfb,0x05,0xc5,
-    0x23,0xb0,0x03,0x4c,0x20,0xcd,0xa4,0x23,
-    0x88,0x8c,0xfb,0x05,0x8a,0x48,0xa2,0x01,
-    0xd0,0x04,0x8a,0x48,0xa2,0x00,0x2c,0x1f,
-    0xc0,0x10,0x05,0xa5,0x21,0x48,0x46,0x21,
-    0x20,0xd1,0xcc,0x2c,0x1f,0xc0,0x10,0x51,
-    0x08,0x78,0xad,0x55,0xc0,0x20,0xd1,0xcc,
-    0xad,0x54,0xc0,0x28,0x68,0x85,0x21,0xd0,
-    0x40,0xbc,0xf9,0xcf,0xb9,0x00,0x00,0xe0,
-    0x01,0xb0,0x02,0xe9,0x00,0x48,0x20,0x54,
-    0xcb,0xa5,0x28,0x85,0x2a,0xa5,0x29,0x85,
-    0x2b,0xa4,0x21,0x88,0x68,0x18,0x7d,0xf0,
-    0xcf,0xd5,0x22,0xf0,0x0d,0x48,0x20,0x54,
-    0xcb,0xb1,0x28,0x91,0x2a,0x88,0x10,0xf9,
-    0x30,0xdf,0xe0,0x00,0xd0,0x0a,0x20,0x54,
-    0xcb,0xb1,0x28,0x91,0x2a,0x88,0x10,0xf9,
-    0x60,0xb4,0x22,0x8a,0xf0,0x01,0x88,0x98,
-    0x20,0x54,0xcb,0x68,0xaa,0x20,0x1a,0xcc,
-    0x4c,0x51,0xcb,0x20,0x48,0xcd,0xad,0xfb,
-    0x05,0x48,0x10,0x06,0x20,0x51,0xcb,0x20,
-    0x1a,0xcc,0xee,0xfb,0x05,0xad,0xfb,0x05,
-    0xc5,0x23,0x90,0xf0,0x68,0x8d,0xfb,0x05,
-    0x10,0xde,0x20,0x0d,0xcc,0x4c,0x23,0xcd,
-    0xac,0x7b,0x05,0x4c,0x54,0xcd,0xa9,0xa0,
-    0x20,0xf2,0xce,0xc8,0xc4,0x21,0x90,0xf6,
-    0x60,0xa9,0x00,0x85,0x20,0x2c,0x1a,0xc0,
-    0x30,0x02,0xa9,0x14,0x85,0x22,0xa9,0x18,
-    0x85,0x23,0xa9,0x28,0x85,0x21,0x2c,0x1f,
-    0xc0,0x10,0x03,0x20,0xdb,0xcd,0x60,0x20,
-    0x24,0xcb,0xd0,0x1e,0x20,0x9b,0xcd,0x2c,
-    0x1a,0xc0,0x30,0x04,0xa9,0x14,0x85,0x22,
-    0x2c,0x18,0xc0,0x30,0x0d,0x4c,0x32,0xce,
-    0xad,0xfb,0x04,0x29,0x20,0xd0,0x03,0x20,
-    0xaa,0xcd,0x60,0xa9,0x00,0x85,0x22,0x85,
-    0x20,0xa9,0x50,0x85,0x21,0xa9,0x18,0x85,
-    0x23,0x60,0xa9,0x00,0x85,0x22,0x85,0x20,
-    0xa9,0x18,0x85,0x23,0xa9,0x28,0x85,0x21,
-    0x2c,0x1f,0xc0,0x10,0x03,0x20,0xdb,0xcd,
-    0xa9,0x17,0x8d,0xfb,0x05,0x20,0x51,0xcb,
-    0xa9,0x00,0x8d,0x7b,0x05,0x8d,0x0e,0xc0,
-    0xa9,0xff,0x8d,0xfb,0x04,0x20,0x93,0xfe,
-    0x4c,0x89,0xfe,0xad,0xfb,0x05,0x48,0xad,
-    0x7b,0x05,0x48,0xa9,0x17,0x85,0x2a,0x8d,
-    0x01,0xc0,0xa5,0x2a,0x20,0x54,0xcb,0x20,
-    0x0a,0xce,0xc6,0x2a,0x30,0x0b,0x2c,0x1a,
-    0xc0,0x30,0xef,0xa5,0x2a,0xc9,0x14,0xb0,
-    0xe9,0x8d,0x00,0xc0,0x8d,0x0c,0xc0,0x4c,
-    0x58,0xce,0x08,0x78,0xa0,0x28,0x84,0x2b,
-    0x2c,0x54,0xc0,0x20,0x22,0xce,0x2c,0x55,
-    0xc0,0x20,0x22,0xce,0xa4,0x2b,0xd0,0xf3,
-    0x28,0x60,0xc6,0x2b,0xa5,0x2b,0x4a,0xa8,
-    0xb1,0x28,0xa4,0x2b,0x2c,0x54,0xc0,0x91,
-    0x28,0x60,0xad,0xfb,0x05,0x48,0xad,0x7b,
-    0x05,0x48,0xa9,0x17,0x85,0x2a,0xa5,0x2a,
-    0x20,0x54,0xcb,0x20,0x63,0xce,0xc6,0x2a,
-    0x30,0x0b,0x2c,0x1a,0xc0,0x30,0xef,0xa5,
-    0x2a,0xc9,0x14,0xb0,0xe9,0x8d,0x0d,0xc0,
-    0x68,0x8d,0x7b,0x05,0x68,0x8d,0xfb,0x05,
-    0x4c,0x51,0xcb,0x08,0x78,0xa0,0x00,0x84,
-    0x2b,0x8c,0x01,0xc0,0x2c,0x54,0xc0,0xb1,
-    0x28,0x2c,0x55,0xc0,0x20,0xa3,0xce,0x2c,
-    0x54,0xc0,0xb1,0x28,0x20,0xa3,0xce,0xc0,
-    0x28,0x90,0xec,0x20,0x91,0xce,0x2c,0x55,
-    0xc0,0x20,0x91,0xce,0x2c,0x54,0xc0,0x28,
-    0x60,0xa0,0x14,0xa9,0xa0,0x24,0x32,0x30,
-    0x02,0x29,0x7f,0x91,0x28,0xc8,0xc0,0x28,
-    0xd0,0xf9,0x60,0x48,0x98,0x4a,0xa8,0x68,
-    0x91,0x28,0xe6,0x2b,0xa4,0x2b,0x60,0x8d,
-    0x7b,0x05,0x85,0x24,0x8d,0x7b,0x04,0x2c,
-    0x1f,0xc0,0x10,0x1d,0xa9,0x00,0x85,0x24,
-    0x8d,0x7b,0x04,0xa5,0x21,0x38,0xed,0x7b,
-    0x05,0xc9,0x08,0xb0,0x0c,0x85,0x24,0xa9,
-    0x28,0x38,0xe5,0x24,0x85,0x24,0x8d,0x7b,
-    0x04,0xad,0x7b,0x05,0x60,0x48,0x98,0x48,
-    0xac,0x7b,0x05,0x20,0x01,0xcf,0x49,0x80,
-    0x2c,0x00,0xcf,0x20,0x06,0xcf,0x68,0xa8,
-    0x68,0x60,0x48,0x24,0x32,0x30,0x02,0x49,
-    0x80,0x2c,0x00,0xcf,0x20,0x06,0xcf,0x68,
-    0x60,0xb8,0x20,0x06,0xcf,0x60,0x84,0x1f,
-    0x48,0xad,0x1f,0xc0,0x10,0x32,0xa5,0x1f,
-    0x4a,0xa8,0x70,0x16,0x08,0x78,0xad,0x55,
-    0xc0,0x90,0x03,0xad,0x54,0xc0,0xb1,0x28,
-    0xa8,0xad,0x54,0xc0,0x28,0x68,0x98,0x48,
-    0x50,0x24,0x68,0x48,0x08,0x78,0x48,0xad,
-    0x55,0xc0,0x90,0x03,0xad,0x54,0xc0,0x68,
-    0x91,0x28,0xad,0x54,0xc0,0x28,0x70,0x0e,
-    0xa4,0x1f,0x70,0x06,0x68,0xb1,0x28,0x48,
-    0x50,0x04,0x68,0x48,0x91,0x28,0x68,0xa4,
-    0x1f,0x60,0x48,0x98,0x48,0xac,0x7b,0x05,
-    0x20,0x01,0xcf,0x8d,0x7b,0x06,0x29,0x80,
-    0x49,0xab,0x4c,0x6e,0xcf,0x48,0x98,0x48,
-    0xac,0x7b,0x05,0xad,0x7b,0x06,0x2c,0x00,
-    0xcf,0x20,0x06,0xcf,0x68,0xa8,0x68,0x60,
-    0x48,0x08,0x78,0xad,0x11,0xc0,0x48,0xae,
-    0x12,0xc0,0xad,0x81,0xc0,0xad,0x81,0xc0,
-    0xa0,0x00,0xa9,0xf8,0x85,0x37,0xa5,0x36,
-    0x48,0xa9,0x00,0x85,0x36,0xb1,0x36,0x91,
-    0x36,0xc8,0xd0,0xf9,0xe6,0x37,0xd0,0xf5,
-    0x68,0x85,0x36,0xa9,0xc3,0x85,0x37,0x68,
-    0x10,0x0f,0x8a,0x10,0x06,0xad,0x80,0xc0,
-    0x4c,0xc5,0xcf,0xad,0x81,0xc0,0x4c,0xc5,
-    0xcf,0x8a,0x10,0x06,0xad,0x88,0xc0,0x4c,
-    0xc5,0xcf,0xad,0x89,0xc0,0x28,0x68,0x60,
-    0xad,0xfb,0x04,0x29,0x01,0xd0,0x03,0x20,
-    0x9b,0xcd,0xa9,0xff,0x85,0x32,0xad,0xfb,
-    0x04,0x29,0x04,0xf0,0x02,0x46,0x32,0xad,
-    0x7b,0x07,0x85,0x28,0xad,0xfb,0x07,0x85,
-    0x29,0x60,0x28,0x42,0x4c,0x7c,0x9b,0xe9,
-    0xff,0x01,0x89,0xe0,0xec,0xcc,0xd2,0xd8,
-    0xe9,0x23,0x22,0xe6,0x00,0x00,0x00,0x00,
     0x6f,0xd8,0x65,0xd7,0xf8,0xdc,0x94,0xd9,
     0xb1,0xdb,0x30,0xf3,0xd8,0xdf,0xe1,0xdb,
     0x8f,0xf3,0x98,0xf3,0xe4,0xf1,0xdd,0xf1,
@@ -1793,7 +1281,7 @@ const rom = new Uint8Array([
     0xa9,0x00,0x85,0x14,0x4c,0xf0,0xd8,0x20,
     0xf8,0xe6,0xca,0x8a,0xc9,0x28,0x90,0x0a,
     0xe9,0x28,0x48,0x20,0xfb,0xda,0x68,0x4c,
-    0xec,0xf7,0x85,0x24,0x60,0xcb,0xd2,0x78,
+    0xec,0xf7,0x85,0x24,0x60,0xcb,0xd2,0xd7,
     0x4a,0x08,0x20,0x47,0xf8,0x28,0xa9,0x0f,
     0x90,0x02,0x69,0xe0,0x85,0x2e,0xb1,0x26,
     0x45,0x30,0x25,0x2e,0x51,0x26,0x91,0x26,
@@ -1869,11 +1357,11 @@ const rom = new Uint8Array([
     0x85,0x45,0x68,0x48,0x0a,0x0a,0x0a,0x30,
     0x03,0x6c,0xfe,0x03,0x28,0x20,0x4c,0xff,
     0x68,0x85,0x3a,0x68,0x85,0x3b,0x6c,0xf0,
-    0x03,0x20,0x82,0xf8,0x20,0xda,0xfa,0x4c,
+    0x03,0x20,0x82,0xf8,0x20,0x2d,0xff,0x4c,
     0x65,0xff,0xd8,0x20,0x84,0xfe,0x20,0x2f,
     0xfb,0x20,0x93,0xfe,0x20,0x89,0xfe,0xad,
-    0x58,0xc0,0xad,0x5a,0xc0,0xa0,0x05,0x20,
-    0xb4,0xfb,0xea,0xad,0xff,0xcf,0x2c,0x10,
+    0x58,0xc0,0xad,0x5a,0xc0,0xad,0x5d,0xc0,
+    0xad,0x5f,0xc0,0xad,0xff,0xcf,0x2c,0x10,
     0xc0,0xd8,0x20,0x3a,0xff,0xad,0xf3,0x03,
     0x49,0xa5,0xcd,0xf4,0x03,0xd0,0x17,0xad,
     0xf2,0x03,0xd0,0x0f,0xa9,0xe0,0xcd,0xf3,
@@ -1884,23 +1372,23 @@ const rom = new Uint8Array([
     0x85,0x01,0xa0,0x07,0xc6,0x01,0xa5,0x01,
     0xc9,0xc0,0xf0,0xd7,0x8d,0xf8,0x07,0xb1,
     0x00,0xd9,0x01,0xfb,0xd0,0xec,0x88,0x88,
-    0x10,0xf5,0x6c,0x00,0x00,0xea,0xea,0x20,
-    0x8e,0xfd,0xa9,0x45,0x85,0x40,0xa9,0x00,
-    0x85,0x41,0xa2,0xfb,0xa9,0xa0,0x20,0xed,
-    0xfd,0xbd,0x1e,0xfa,0x20,0xed,0xfd,0xa9,
-    0xbd,0x20,0xed,0xfd,0xb5,0x4a,0x20,0xda,
-    0xfd,0xe8,0x30,0xe8,0x60,0x59,0xfa,0x00,
+    0x10,0xf5,0x6c,0x00,0x00,0xea,0xea,0x60,
+    0x48,0xa9,0x0f,0x25,0x4f,0xd0,0x0c,0x68,
+    0x48,0xd1,0x28,0xd0,0x02,0xa9,0xdf,0x91,
+    0x28,0xe6,0x4f,0x68,0x2c,0x00,0xc0,0x60,
+    0xad,0x56,0xc0,0xad,0x53,0xc0,0x60,0xea,
+    0xea,0xea,0xea,0xea,0xea,0x59,0xfa,0x00,
     0xe0,0x45,0x20,0xff,0x00,0xff,0x03,0xff,
-    0x3c,0xc1,0xf0,0xf0,0xec,0xe5,0xa0,0xdd,
+    0x3c,0xc1,0xd0,0xd0,0xcc,0xc5,0xa0,0xdd,
     0xdb,0xc4,0xc2,0xc1,0xff,0xc3,0xff,0xff,
     0xff,0xc1,0xd8,0xd9,0xd0,0xd3,0xad,0x70,
     0xc0,0xa0,0x00,0xea,0xea,0xbd,0x64,0xc0,
     0x10,0x04,0xc8,0xd0,0xf8,0x88,0x60,0xa9,
     0x00,0x85,0x48,0xad,0x56,0xc0,0xad,0x54,
     0xc0,0xad,0x51,0xc0,0xa9,0x00,0xf0,0x0b,
-    0xad,0x50,0xc0,0xad,0x53,0xc0,0x20,0x36,
+    0xad,0x50,0xc0,0x20,0xf0,0xfa,0x20,0x36,
     0xf8,0xa9,0x14,0x85,0x22,0xa9,0x00,0x85,
-    0x20,0xa0,0x08,0xd0,0x5f,0xa9,0x18,0x85,
+    0x20,0xa9,0x28,0x85,0x21,0xa9,0x18,0x85,
     0x23,0xa9,0x17,0x85,0x25,0x4c,0x22,0xfc,
     0x20,0x58,0xfc,0xa0,0x08,0xb9,0x08,0xfb,
     0x99,0x0e,0x04,0x88,0xd0,0xf7,0x60,0xad,
@@ -1910,10 +1398,10 @@ const rom = new Uint8Array([
     0xac,0x00,0xc0,0x10,0xfb,0xc0,0x83,0xf0,
     0x03,0x2c,0x10,0xc0,0x4c,0xfd,0xfb,0x38,
     0x4c,0x2c,0xfc,0xa8,0xb9,0x48,0xfa,0x20,
-    0x97,0xfb,0x20,0x21,0xfd,0xc9,0xce,0xb0,
+    0x97,0xfb,0x20,0x0c,0xfd,0xc9,0xce,0xb0,
     0xee,0xc9,0xc9,0x90,0xea,0xc9,0xcc,0xf0,
-    0xe6,0xd0,0xe8,0x06,0x08,0x78,0x2c,0x15,
-    0xc0,0x08,0x8d,0x07,0xc0,0x4c,0x00,0xc1,
+    0xe6,0xd0,0xe8,0xc9,0x94,0xd0,0xc1,0xa5,
+    0x32,0x49,0xc0,0x4c,0x46,0xfd,0xea,0xea,
     0xea,0x48,0x4a,0x29,0x03,0x09,0x04,0x85,
     0x29,0x68,0x29,0x18,0x90,0x02,0x69,0x7f,
     0x85,0x28,0x0a,0x0a,0x05,0x28,0x85,0x28,
@@ -1930,19 +1418,19 @@ const rom = new Uint8Array([
     0x20,0x85,0x28,0x60,0x49,0xc0,0xf0,0x28,
     0x69,0xfd,0x90,0xc0,0xf0,0xda,0x69,0xfd,
     0x90,0x2c,0xf0,0xde,0x69,0xfd,0x90,0x5c,
-    0xd0,0xe9,0xa0,0x00,0xf0,0x2c,0xa8,0xc3,
-    0xa9,0xa0,0xb1,0xb9,0xb8,0xb1,0xad,0xb8,
-    0xb2,0xac,0xa0,0xc1,0xd0,0xd0,0xcc,0xc5,
-    0xa0,0x01,0xd0,0x16,0xd2,0xc9,0xc3,0xcb,
-    0xa0,0xc1,0xa9,0x00,0x85,0x24,0xe6,0x25,
+    0xd0,0xe9,0xa4,0x24,0xa5,0x25,0x48,0x20,
+    0x24,0xfc,0x20,0x9e,0xfc,0xa0,0x00,0x68,
+    0x69,0x00,0xc5,0x23,0x90,0xf0,0xb0,0xca,
+    0xa5,0x22,0x85,0x25,0xa0,0x00,0x84,0x24,
+    0xf0,0xe4,0xa9,0x00,0x85,0x24,0xe6,0x25,
     0xa5,0x25,0xc5,0x23,0x90,0xb6,0xc6,0x25,
-    0xa0,0x02,0x4c,0xb4,0xfb,0x48,0xad,0x18,
-    0xc0,0x0a,0x68,0x2c,0x1c,0xc0,0x08,0x90,
-    0x03,0x8d,0x54,0xc0,0x2c,0x15,0xc0,0x8d,
-    0x06,0xc0,0x58,0x78,0x10,0x03,0x8d,0x07,
-    0xc0,0x28,0x90,0x05,0x10,0x03,0x2c,0x55,
-    0xc0,0x60,0xea,0xea,0x18,0xb0,0x38,0x84,
-    0x1f,0xa0,0x03,0x90,0xcd,0xc8,0xd0,0xca,
+    0xa5,0x22,0x48,0x20,0x24,0xfc,0xa5,0x28,
+    0x85,0x2a,0xa5,0x29,0x85,0x2b,0xa4,0x21,
+    0x88,0x68,0x69,0x01,0xc5,0x23,0xb0,0x0d,
+    0x48,0x20,0x24,0xfc,0xb1,0x28,0x91,0x2a,
+    0x88,0x10,0xf9,0x30,0xe1,0xa0,0x00,0x20,
+    0x9e,0xfc,0xb0,0x86,0xa4,0x24,0xa9,0xa0,
+    0x91,0x28,0xc8,0xc4,0x21,0x90,0xf9,0x60,
     0x38,0x48,0xe9,0x01,0xd0,0xfc,0x68,0xe9,
     0x01,0xd0,0xf6,0x60,0xe6,0x42,0xd0,0x02,
     0xe6,0x43,0xa5,0x3c,0xc5,0x3e,0xa5,0x3d,
@@ -1956,21 +1444,21 @@ const rom = new Uint8Array([
     0xf5,0x60,0x20,0xfd,0xfc,0x88,0xad,0x60,
     0xc0,0x45,0x2f,0x10,0xf8,0x45,0x2f,0x85,
     0x2f,0xc0,0x80,0x60,0xa4,0x24,0xb1,0x28,
-    0x48,0x29,0x3f,0x09,0x40,0x91,0x28,0x68,
-    0x6c,0x38,0x00,0xa0,0x06,0x4c,0xb4,0xfb,
-    0xea,0x20,0x0c,0xfd,0xa0,0x07,0x4c,0xb4,
-    0xfb,0x8d,0x06,0xc0,0x28,0x60,0x60,0x20,
-    0x21,0xfd,0x20,0xa5,0xfb,0x20,0x0c,0xfd,
-    0xc9,0x9b,0xf0,0xf3,0x60,0xa5,0x32,0x48,
-    0xa9,0xff,0xea,0xea,0xbd,0x00,0x02,0x20,
-    0xed,0xfd,0x68,0x85,0x32,0xbd,0x00,0x02,
+    0x48,0x29,0x3f,0x09,0x40,0xea,0xea,0x68,
+    0x6c,0x38,0x00,0xe6,0x4e,0xd0,0x02,0xe6,
+    0x4f,0x20,0xd8,0xfa,0x10,0xf5,0x91,0x28,
+    0xad,0x00,0xc0,0x2c,0x10,0xc0,0x60,0x20,
+    0x0c,0xfd,0x20,0xa5,0xfb,0x20,0x0c,0xfd,
+    0xc9,0x9b,0xf0,0xf3,0x60,0xbd,0x00,0x02,
+    0x20,0xed,0xfd,0x4c,0x4d,0xfd,0x85,0x32,
+    0xa9,0x94,0x4c,0x78,0xfb,0xbd,0x00,0x02,
     0xc9,0x88,0xf0,0x1d,0xc9,0x98,0xf0,0x0a,
     0xe0,0xf8,0x90,0x03,0x20,0x3a,0xff,0xe8,
     0xd0,0x13,0xa9,0xdc,0x20,0xed,0xfd,0x20,
     0x8e,0xfd,0xa5,0x33,0x20,0xed,0xfd,0xa2,
     0x01,0x8a,0xf0,0xf3,0xca,0x20,0x35,0xfd,
     0xc9,0x95,0xd0,0x02,0xb1,0x28,0xc9,0xe0,
-    0x90,0x02,0x29,0xff,0x9d,0x00,0x02,0xc9,
+    0x90,0x02,0x29,0xdf,0x9d,0x00,0x02,0xc9,
     0x8d,0xd0,0xb2,0x20,0x9c,0xfc,0xa9,0x8d,
     0xd0,0x5b,0xa4,0x3d,0xa6,0x3c,0x20,0x8e,
     0xfd,0x20,0x40,0xf9,0xa0,0x00,0xa9,0xad,
@@ -1985,7 +1473,7 @@ const rom = new Uint8Array([
     0xe5,0xfd,0x68,0x29,0x0f,0x09,0xb0,0xc9,
     0xba,0x90,0x02,0x69,0x06,0x6c,0x36,0x00,
     0xc9,0xa0,0x90,0x02,0x25,0x32,0x84,0x35,
-    0x48,0x20,0x78,0xfb,0x68,0xa4,0x35,0x60,
+    0x48,0x20,0xb3,0xfb,0x68,0xa4,0x35,0x60,
     0xc6,0x34,0xf0,0x9f,0xca,0xd0,0x16,0xc9,
     0xba,0xd0,0xbb,0x85,0x31,0xa5,0x3e,0x91,
     0x40,0xe6,0x40,0xd0,0x02,0xe6,0x41,0x60,
@@ -2007,11 +1495,11 @@ const rom = new Uint8Array([
     0x1b,0xd0,0x08,0xa9,0x00,0x85,0x3e,0xa2,
     0x36,0xa0,0xf0,0xa5,0x3e,0x29,0x0f,0xf0,
     0x06,0x09,0xc0,0xa0,0x00,0xf0,0x02,0xa9,
-    0xfd,0x94,0x00,0x95,0x01,0x60,0xea,0xd1,
+    0xfd,0x94,0x00,0x95,0x01,0x60,0xea,0xea,
     0x4c,0x00,0xe0,0x4c,0x03,0xe0,0x20,0x75,
-    0xfe,0x20,0x3f,0xff,0x6c,0x3a,0x00,0x4c,
-    0xd7,0xfa,0x60,0xea,0x60,0xc2,0xf2,0xf9,
-    0xe1,0xee,0x4c,0xf8,0x03,0xa9,0x40,0x20,
+    0xfe,0x20,0x3f,0xff,0x6c,0x3a,0x00,0x60,
+    0xea,0xea,0x60,0xea,0x60,0xea,0xea,0xea,
+    0xea,0xea,0x4c,0xf8,0x03,0xa9,0x40,0x20,
     0xc9,0xfc,0xa0,0x27,0xa2,0x00,0x41,0x3c,
     0x48,0xa1,0x3c,0x20,0xed,0xfe,0x20,0xba,
     0xfc,0xa0,0x1d,0x68,0x90,0xee,0xa0,0x22,
@@ -2049,11 +1537,11 @@ const rom = new Uint8Array([
     0xa7,0xc6,0x99,0xb2,0xc9,0xbe,0xc1,0x35,
     0x8c,0xc4,0x96,0xaf,0x17,0x17,0x2b,0x1f,
     0x83,0x7f,0x5d,0xcc,0xb5,0xfc,0x17,0x17,
-    0xf5,0x03,0xfb,0x03,0x62,0xfa,0x40,0xfa
+    0xf5,0x03,0xfb,0x03,0x62,0xfa,0x40,0xfa,
 ]) as ReadonlyUint8Array;
 
-export default class Apple2eROM extends ROM {
+export default class Apple2jROM extends ROM {
     constructor() {
-        super(0xc0, 0xff, rom);
+        super(0xd0, 0xff, rom);
     }
 }
