+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-      "module": "commonjs",
+      "module": "esnext",
       "esModuleInterop": true,
       "allowSyntheticDefaultImports": true,
       "target": "es6",
